- Оновлено модель Ingredient: змінено тип _id з STRING на INTEGER з автоінкрементом. - Додано ініціалізацію моделей у seed.js
