--- db/models/Ingredient.js
@@ -1,35 +1,0 @@
-import { DataTypes } from "sequelize";
-import sequelize from "../Sequelize.js";
-
-const Ingredient = sequelize.define(
-    "ingredient",
-    {
-        _id: {
-            type: DataTypes.STRING,
-            primaryKey: true,
-            allowNull: false,
-        },
-        name: {
-            type: DataTypes.STRING,
-            allowNull: false,
-            unique: true,
-        },
-        desc: {
-            type: DataTypes.TEXT,
-            allowNull: true,
-        },
-        img: {
-            type: DataTypes.STRING,
-            allowNull: true,
-        },
-    },
-    {
-        autoIncrement: false,
-        timestamps: true,
-    }
-);
-
-// Ingredient.sync();
-// Ingredient.sync({ force: true });
-
-export default Ingredient;
